Fix RPM build spec file creation
For RPM build we were getting the issue while craeting teh spec files. Issue is when we pass the channels in a loop after reading the template, for the first pass of template it reads file and update teh content but for further channel file read reached EOF. So it had no content to read. To fix this, we need to reset the file pointer to the beginning of file.